SDV & CyberSecurity/requirements.txt
# CAN Bus IDS System Requirements

# Core dependencies
python-can>=4.0.0
pandas>=1.5.0
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0

# Enhanced ML dependencies
xgboost>=1.7.0
scipy>=1.10.0

# Dashboard dependencies
streamlit>=1.28.0
plotly>=5.15.0

# Optional: for better CAN interface support
# cantools>=36.0.0
# python-can[serial]>=4.0.0

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0